backend/pyproject.toml
[project]
name = "llm-eval-pipeline"
version = "0.1.0"
description = "Production-grade LLM Evaluation & Red-Teaming Pipeline"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Hemant BK" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "redis[hiredis]>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "prometheus-client>=0.21.0",
    "tenacity>=9.0.0",
    "pybreaker>=1.2.0",
    "pydantic-settings>=2.5.0",
    "alembic>=1.13.0",
    "structlog>=24.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
    "locust>=2.29.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "coverage>=7.6.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
